// src/types/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Date custom scalar type */
  Date: any;
};
// ...
export type Category = {
  __typename?: 'Category';
  _id: Scalars['ID'];
  assets: Array<Maybe<Scalars['String']>>;
  children: Array<Maybe<Category>>;
  createdAt: Scalars['Date'];
  description?: Maybe<Scalars['String']>;
  name: Scalars['String'];
  parent?: Maybe<Category>;
  products: Array<Product>;
  updatedAt: Scalars['Date'];
};

export type Conditionals = {
  __typename?: 'Conditionals';
  hasImages: Scalars['Boolean'];
  isActive: Scalars['Boolean'];
  isInventoryManaged: Scalars['Boolean'];
  isSoldOut: Scalars['Boolean'];
};

export type Discount = {
  __typename?: 'Discount';
  _id: Scalars['String'];
  code: Scalars['String'];
  currentQuantity: Scalars['Int'];
  endsOn?: Maybe<Scalars['Date']>;
  isExpired: Scalars['Boolean'];
  limitQuantity?: Maybe<Scalars['Int']>;
  products: Array<Product>;
  startsOn: Scalars['Date'];
  type: DiscountType;
  value: Scalars['Int'];
};

export enum DiscountType {
  Number = 'number',
  Percentage = 'percentage'
}

export type Inventory = {
  __typename?: 'Inventory';
  avaible?: Maybe<Scalars['Int']>;
  managed: Scalars['Boolean'];
};
// ...
export type Product = {
  __typename?: 'Product';
  _id: Scalars['ID'];
  assets: Array<Maybe<Scalars['String']>>;
  categories: Array<Maybe<Category>>;
  conditionals: Conditionals;
  createdAt: Scalars['Date'];
  description?: Maybe<Scalars['String']>;
  discount?: Maybe<Discount>;
  image?: Maybe<Scalars['String']>;
  inventory: Inventory;
  name: Scalars['String'];
  price: Scalars['Float'];
  relatedProducts: Array<Maybe<Product>>;
  sales: Scalars['Int'];
  shippingMethods: Array<Maybe<ShippingMethod>>;
  updatedAt: Scalars['Date'];
  variantGroups: Array<Maybe<VariantGroup>>;
  variants: Array<Maybe<Variant>>;
};
// ...
export type ShippingMethod = {
  __typename?: 'ShippingMethod';
  _id: Scalars['ID'];
  name: Scalars['String'];
  price: Scalars['Float'];
  zones: Array<ShippingMethodZone>;
};

export type ShippingMethodZone = {
  __typename?: 'ShippingMethodZone';
  country: Scalars['String'];
  regions: Array<Scalars['String']>;
};

export type User = {
  __typename?: 'User';
  _id: Scalars['ID'];
  email: Scalars['String'];
  firstName: Scalars['String'];
  image?: Maybe<Scalars['String']>;
  lastName: Scalars['String'];
  password?: Maybe<Scalars['String']>;
  provider: Scalars['String'];
  type: Scalars['String'];
};

export type Variant = {
  __typename?: 'Variant';
  _id: Scalars['String'];
  assets?: Maybe<Array<Maybe<Scalars['String']>>>;
  createdAt: Scalars['Date'];
  description?: Maybe<Scalars['String']>;
  inventory?: Maybe<Scalars['Int']>;
  options: Array<VariantOption>;
  price: Scalars['Float'];
  productId: Scalars['ID'];
  updatedAt: Scalars['Date'];
};

export type VariantGroup = {
  __typename?: 'VariantGroup';
  _id: Scalars['ID'];
  assets?: Maybe<Array<Maybe<Scalars['String']>>>;
  createdAt: Scalars['Date'];
  name: Scalars['String'];
  options: Array<Scalars['String']>;
  productId: Scalars['String'];
  updatedAt: Scalars['Date'];
};

export type VariantOption = {
  __typename?: 'VariantOption';
  option: Scalars['String'];
  variantGroup: Scalars['String'];
};